feat: soporte de fracciones en stock (Float) para productos por KG
Cambia stock, minStock y delta de Int a Float en schema y DB.
Actualiza validadores Zod e inputs UI para permitir cantidades decimales.

+++ src/app/(app)/productos/[id]/actions.ts
@@ -7,7 +7,7 @@ import { sendLowStockAlertWithRetry } from '@/lib/email'
 
 const MovementSchema = z.object({
   productId: z.string().uuid(),
-  delta: z.coerce.number().int().refine((n) => n !== 0, { message: 'La cantidad debe ser distinta de cero.' }),
+  delta: z.coerce.number().refine((n) => n !== 0, { message: 'La cantidad debe ser distinta de cero.' }),
   motivo: z.string().max(500).optional().nullable(),
 })
